+++ src/blob/handlers/BlobHandler.ts
@@ -863,7 +863,8 @@ export default class BlobHandler extends BaseHandler implements IBlobHandler {
     // Deserializer doesn't handle range header currently, manually parse range headers here
     const rangesParts = deserializePageBlobRangeHeader(
       context.request!.getHeader("range"),
-      context.request!.getHeader("x-ms-range")
+      context.request!.getHeader("x-ms-range"),
+      false
     );
     const rangeStart = rangesParts[0];
     let rangeEnd = rangesParts[1];
@@ -929,6 +930,15 @@ export default class BlobHandler extends BaseHandler implements IBlobHandler {
       body = await bodyGetter();
     }
 
+    let contentRange: string | undefined;
+    if (
+      context.request!.getHeader("range") ||
+      context.request!.getHeader("x-ms-range")
+    ) {
+      contentRange = `bytes ${rangeStart}-${rangeEnd}/${blob.properties
+        .contentLength!}`;
+    }
+
     const response: Models.BlobDownloadResponse = {
       statusCode: rangesParts[1] === Infinity ? 200 : 206,
       body,
@@ -939,6 +949,7 @@ export default class BlobHandler extends BaseHandler implements IBlobHandler {
       version: BLOB_API_VERSION,
       ...blob.properties,
       contentLength,
+      contentRange,
       contentMD5,
       blobContentMD5: blob.properties.contentMD5,
       isServerEncrypted: true,