+++ ChangeLog.md
@@ -25,6 +25,7 @@ Blob:
 - Fixed a bug that Azurite Blob service cannot start in Mac as Visual Studio Extension.
 - Fixed a bug that persistency location cannot be customized through -l parameter.
 - Fixed a bug that GC will remove uncommitted blocks.
+- Fixed a bug that download page blob doesn't return content range header.
 - Fixed a bug that uncommiited block blob invalid length.
 
 Queue: